backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
pdfplumber
python-docx
groq
sqlalchemy
python-multipart
passlib[bcrypt]
python-jose[cryptography]
